String replacement
I had to replace some utf codes that were still lingering in the rendered data (questions and options). This isn't meant to be so but I don't have access to whoever created this data. So I had to work around this on the client side. I'll definitely still pay attention for more bugs.